docker-guide/05-real-world-projects/full-stack-todo-app/backend/requirements.txt
Flask==2.0.2
psycopg2-binary==2.9.1 # For PostgreSQL
